AG-12563 - Move util file
