+++ packages/ag-charts-enterprise/src/features/annotations/parallel-channel/parallelChannelProperties.ts
@@ -12,7 +12,7 @@ import {
     Stroke,
 } from '../annotationProperties';
 import { type AnnotationContext, type AnnotationOptionsColorPickerType, AnnotationType } from '../annotationTypes';
-import { getLineCap, getLineDash } from '../properties/util';
+import { getLineCap, getLineDash } from '../utils/line';
 import { validateDatumLine } from '../utils/validation';
 
 const { NUMBER, STRING, OBJECT, BaseProperties, Validate, isObject } = _ModuleSupport;
