+++ packages/ag-charts-enterprise/src/features/annotations/utils/line.ts
